Conflicts:
	src/main/java/com/onarandombox/MultiverseCore/MultiverseCore.java
	src/main/java/com/onarandombox/MultiverseCore/MultiverseCoreConfiguration.java
	src/main/java/com/onarandombox/MultiverseCore/commands/ConfigCommand.java
	src/main/java/com/onarandombox/MultiverseCore/utils/MVPermissions.java
	src/main/resources/defaults/config.yml

+++ src/main/java/com/onarandombox/MultiverseCore/api/MultiverseCoreConfig.java
@@ -48,7 +48,7 @@ public interface MultiverseCoreConfig extends ConfigurationSerializable {
      * Gets version.
      * @return version.
      */
-    int getVersion();
+    double getVersion();
 
     /**
      * Sets messageCooldown.
